pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
djangorestframework = "^3.13.1"
drf-yasg = "^1.21.3"
django-debug-toolbar = "^3.6.0"
django-oauth-toolkit = "^2.1.0"
django-redis = "^5.2.0"
celery = "^5.2.7"
django-two-factor-auth = {extras = ["phonenumbers"], version = "^1.14.0"}
bandit = "^1.7.4"
pylama = "^8.4.1"
black = "^22.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
